Increase default transientstoreMaxBlockRetention
transientstoreMaxBlockRetention is used to trigger the purge of private
data in the transient store.

The current default value is 1000, which is OK for simple cases with
several blocks per second. Howerver, in our test with 10K tps case, it
is possible to purge the private data too early before the peer gets a
chance to commit the tx into the ledger.

With a larger retention window, the only cost is a larger transient
database in the peer node, which is not a big deal.

Hence, this patchset suggest to increase the default value to a larger
value of 100000, and it is open to have other reasonable value too.

Change-Id: I24eb6a6c4be38fbb964970b518d40c72e02e12ff
Signed-off-by: Baohua Yang <yangbaohua@gmail.com>
Signed-off-by: Baohua Yang <baohua.yang@oracle.com>