Detect your own endpoint in HeightsByEndpoints()
Currently, when an orderer node chooses a single node endpoint to pull blocks from (and the rest to pull headers from),
it may pick itself, which is a problem because this endpoint will never bring it any blocks since it is itself.

This commit makes the function HeightsByEndpoint return the endpoint of yourself, such that the caller can filter it out.

The identification of your own endpoint is done by looking at the TLS certificate of the remote node and comparing it with your own.

Signed-off-by: Yacov Manevich <yacov.manevich@ibm.com>