BFT: skip self-endpoint when pulling blocks (#4693)
The orderer source provides a set of orderers to connect to, and is used by the BFTDeliverer.
The BFTDeliverer is employed in the peer and in the orderer. When used in the orderer, it should
skip the self-endpoint. In this commit we allow the orderer source to be set with the self-endpoint,
and thus allow it to skip it when providing a set of orderers to pull blocks from.


Change-Id: Ia183eef4f263943d5f387e8156d450790bee5cb2

Signed-off-by: Yoav Tock <tock@il.ibm.com>