Fix test flake in TestImpatientStreamDetectSelf
The implementation of the fake/mock orderer in the unit tests
reads a seek envelope and the test fails if reading fails.

When the TestImpatientStreamDetectSelf test ends, it shuts down the fake orderer,
and as a result, reading from the stream fails and the test fails also.

This commit simply short-circuits the test to avoid reading from the stream,
but planting an error before establishing the stream, as reading from the stream
is not needed for the test but only establishing it in order to get the TLS certificate.

Signed-off-by: Yacov Manevich <yacov.manevich@ibm.com>