Set leader rotation disabled but make it possible later
The current code sets leader election disabled by hardcoding it.
More specifically, it overrides leader rotation in the code section that parses the config.
This means that if we choose to make it possible to enable it at a later version, it won't be
possible to run mixed versions of v3.0.

This commit makes sure that if we decide to permit leader rotation later on,
it will be possible if the majority of nodes desire so.

It achieves this by disabling leader rotation in two places:
- Channel creation
- Validation of config update transactions

However, the config parsing logic now honors leader election if it's enabled in the channel config.

Signed-off-by: Yacov Manevich <yacov.manevich@ibm.com>