Nominate Fedor Partanskiy (@pfi79) to the maintainer
@pfi79, has demonstrate very active contribution over the past year,
assiting with code review, fix and providing patches, maintainer
unit-tests aiding to reduce amount of flakes. Was activelly taking
care to align the project with up to date libraries and go eco-system
updates.

In addition to their technical contributions, Fedor has shown a clear
commitment to the project’s success and growth. He have taken the
initiative to address issues promptly and have actively participated
in discussions to maintain the project’s direction. His consistent
effort and dedication make him well-suited for the responsibilities
of a maintainer, ensuring the project’s continued success and reliability.

Signed-off-by: Artem <artem@bargr.net>