Remove attestation server
The attestation server was a concept that was abandoned during the implementation of the BFT block puller.
It is now dead code.

The aim of the attestation server was to pull only headers from the delivery server. It was implemented in a simpler fashion by augmenting the
`seekInfo.ContentType` with `ab.SeekInfo_HEADER_WITH_SIG`
which causes the desired behavior (see common/deliver/deliver.go L338).

Signed-off-by: Yoav Tock <tock@il.ibm.com>