Run vulnerability scan on latest release tags
Vulnerability scans were previously run on the latest state of currently
developed branches. This provided assurance that the current branch
state did not contain known vulnerabilities in dependencies, but did not
provide assurance that the currently released code was free of
vulnerabilities.

This change runs vulnerability scans on the most recent release version
tag for currently developed branches. Scan failures now indicate that a
new release is required to address vulnerabilities in dependencies.

Signed-off-by: Mark S. Lewis <Mark.S.Lewis@outlook.com>