Use Go version from go.mod in basic-checks
The expected Go version was hard-coded in the Makefile, and this needed
to be manually kept in sync with the CI and go.mod files. Since the
go.mod entry needs to be accurate, this change uses the version from the
go.mod file in the Makefile and no longer requires the Makefile to be
manually kept up-to-date.

Signed-off-by: Mark S. Lewis <Mark.S.Lewis@outlook.com>