Add check for Go module version matching CI (#5153)
The Go directive in the go.mod file is used by vulnerability scanning to
determine the Go standard library version to scan for vulnerabilities.
This needs to match exactly the Go version used in CI to build release
binaries for vulnerability scanning to provide an accurate results for a
given Fabric release.

This change adds a CI check that the Go version used exactly matches the
Go directive in the go.mod file.

Signed-off-by: Mark S. Lewis <Mark.S.Lewis@outlook.com>