Fix rpc test issue (#5291)
Currently in the rpc test, a shared fakeStream is used among multiple
runs. This will accumulate the Send data and make the test failed with
the latest github.com/stretchr/testify v1.11.1 version.

The issue is ignored in the current testify v1.11.0 version due to an
incomplete validation.

The patchset fixes the issue by creating a new fakeStream in each run.

Change-Id: Icab00394dd7635fb552379cbfdf1fe111087b7db

Signed-off-by: Baohua Yang <yangbaohua@gmail.com>